prod-dl-iloan-reporting/unzip-iloan-prd/requirements.txt
attrs==22.1.0
Authlib==1.1.0
bcrypt==4.0.1
cached-property==1.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
elementpath==3.0.2
Flask==2.2.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
paramiko==2.11.0
platformdirs==2.5.2
psycopg2-binary==2.9.4
pycparser==2.21
PyNaCl==1.5.0
pysftp==0.2.9
PySocks==1.7.1
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.10.0
simple-salesforce==1.12.2
urllib3==1.26.12
Werkzeug==2.2.2
zeep==4.1.0
zipp==3.9.0
